Renames our class
In Java, there cannot be two classes in the same scope with the same name.

Because we are importing CSVReader from opencsv, we cannot also name our class CSVReader (without jumping through some hoops anyway).

Otherwise the compiler cannot dissern on line 15, when we create a `new CSVReader`, are we trying to create the one from OpenCSV? Or are we trying to create a new instance of this very class we're working on?